Improved create/patch operations

+++ frictionless/server/helpers/text.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional
+from ...exception import FrictionlessException
 from ...resources import TextResource
 
 if TYPE_CHECKING:
@@ -16,10 +17,19 @@ def read_text(project: Project, *, path: str):
     return text
 
 
-def write_text(project: Project, *, path: str, text: str):
+def write_text(
+    project: Project,
+    *,
+    path: str,
+    text: str,
+    overwrite: Optional[bool] = None,
+    deduplicate: Optional[bool] = None,
+):
     fs = project.filesystem
 
     fullpath = fs.get_fullpath(path)
+    if not overwrite and fullpath.exists():
+        raise FrictionlessException("text already exists")
     source = TextResource(data=text)
     target = TextResource(path=str(fullpath))
     source.write_text(target)
